+++ lib/firebase.ts
@@ -29,14 +29,30 @@ export function docToJSON(doc: DocumentSnapshot) {
     }
 }
 
+export async function fetchMasterList(urlUsername: string) {
+    const ref = collection(getFirestore(), 'users');
+    const userInfoQuery = query(
+        // @ts-ignore
+        ref,
+        where('username', "==", urlUsername)
+    )
+
+    // @ts-ignore
+    const userInfo = docToJSON((await getDocs(userInfoQuery)).docs[0]);
+
+    console.log("DB READ!")
+
+    return userInfo.listedMovies;
+}
+
 
 export async function fetchUserInfo(userId: string) {
-    console.log("read")
+    console.log("DB READ!")
     return getDoc(doc(firestore,"users", userId))
 
 }
 export async function deleteMovieFromList(id: number, list: string, userId: string) {
-
+    console.log("DB READ!")
     fetchUserInfo(userId).then((res) => {
         var listedMovies: ListedMovie[] = res.get("listedMovies")
 
@@ -51,6 +67,7 @@ export async function deleteMovieFromList(id: number, list: string, userId: stri
 }
 
 export async function fetchMovieDetailsForList(movieIds: [number]) {
+    console.log("DB READ!")
     const ref = collection(getFirestore(), 'movies');
     const searchQuery = query(
         // @ts-ignore
@@ -62,12 +79,12 @@ export async function fetchMovieDetailsForList(movieIds: [number]) {
 
 
     // console.table(userInfo);
-    console.log("read")
     // @ts-ignore
     return (await getDocs(searchQuery)).docs.map(docToJSON);
 }
 
 export async function getMovieById(id : number) {
+    console.log("DB READ!")
     const ref = collection(getFirestore(), 'movies');
     const genreQuery = query(
         // @ts-ignore
